ProductsCrud/Controllers/UserController.cs
using Business.Interfaces.Services;
using Business.Models;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Http;
using Microsoft.AspNetCore.Mvc;
using Microsoft.Extensions.Hosting;
using Microsoft.IdentityModel.Tokens;
using System.IdentityModel.Tokens.Jwt;
using System.Security.Claims;
using System.Text;

namespace ProductsCrud.Controllers
{
    [AllowAnonymous]
    public class UserController : Controller
    {
        private readonly IUserService _userService;

        public UserController(IUserService userService)
        {
            _userService = userService;
        }

        public IActionResult CreateUser()
        {
            return View();
        }

        [HttpPost]
        public IActionResult CreateUser(User user)
        {
            if (ModelState.IsValid)
            {
                _userService.CreateUser(user);
                GenerateJwtToken(user.Username);
                return RedirectToAction("Index", "Product");
            }

            return View(user);
        }

        public IActionResult Login()
        {
            return View();
        }

        [HttpPost]
        public IActionResult Login(User user)
        {
            if (_userService.ValidateUser(user))
            {
                GenerateJwtToken(user.Username);
                return RedirectToAction("Index", "Product");
            }

            ModelState.AddModelError(string.Empty, "Invalid username or password");
            return View(user);
        }

        private void GenerateJwtToken(string username)
        {
            var claims = new[]
             {
                new Claim(ClaimTypes.Name, username),
            };

            var token = new JwtSecurityToken(
                issuer: "issuer",
                audience: "audience",
                claims: claims,
                expires: DateTime.UtcNow.AddMinutes(30),
                signingCredentials: new SigningCredentials(new SymmetricSecurityKey(System.Text.Encoding.UTF8.GetBytes("this is my custom Secret key for authentication")), SecurityAlgorithms.HmacSha256)
                
            );
            var JWToken = new JwtSecurityTokenHandler().WriteToken(token);
            //HttpContext.Session.SetString("JWToken", JWToken);
        }

    }
}
